+++ src/compiler/command_line_parser/mod.rs
@@ -1,8 +1,12 @@
 mod lines_0_1000;
 pub use lines_0_1000::*;
-mod lines_1000_2000;
-pub use lines_1000_2000::*;
-mod lines_2000_3000;
-pub use lines_2000_3000::*;
+mod lines_1000_1500;
+pub use lines_1000_1500::*;
+mod lines_1500_2000;
+pub use lines_1500_2000::*;
+mod lines_2000_2500;
+pub use lines_2000_2500::*;
+mod lines_2500_3000;
+pub use lines_2500_3000::*;
 mod lines_3000_end;
 pub use lines_3000_end::*;
